# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: InterHive API
  description: InterHive - Industry Readiness & Talent Connect Platform API
  version: 1.0.0
  contact:
    name: InterHive Team
    email: support@interhive.in
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.interhive.in/api/v1
    description: Production Server
  - url: https://staging-api.interhive.in/api/v1
    description: Staging Server
  - url: http://localhost:3000/api/v1
    description: Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        timestamp:
          type: string
          format: date-time
...
    Company:
      type: object
      required:
        - companyInfo
        - contact
      properties:
        companyInfo:
          type: object
          properties:
            name:
              type: string
            legalName:
              type: string
            registrationNumber:
              type: string
            industry:
              type: array
              items:
                type: string
            size:
              type: integer
            foundedYear:
              type: integer
            website:
              type: string
            description:
              type: string
            logo:
              type: string
        contact:
          type: object
          properties:
            primaryContact:
              type: object
              properties:
                email:
                  type: string
                phone:
                  type: string
                address:
                  type: object
                  properties:
                    street:
                      type: string
                    city:
                      type: string
                    state:
                      type: string
                    country:
                      type: string
                    zipCode:
                      type: string

    # Assessment
...
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            message: Resource not found
            timestamp: '2024-01-01T00:00:00.000Z'
...
paths:
...
  /companies:
    post:
      tags: [Companies]
      summary: Create company profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Company'
      responses:
        201:
          description: Company created

    get:
      tags: [Companies]
      summary: Get all companies
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
        - name: industry
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, verified, active, suspended, inactive]
      responses:
        200:
          description: Companies retrieved
  /companies/{id}:
    get:
      tags: [Companies]
      summary: Get company by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Company retrieved
        404:
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [Companies]
      summary: Update company
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Company'
      responses:
        200:
          description: Company updated
        404:
          $ref: '#/components/responses/NotFoundError'
  /companies/{id}/requirements:
    post:
      tags: [Companies]
      summary: Create talent requirement
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - position
                - count
                - skills
              properties:
                position:
                  type: string
                department:
                  type: string
                count:
                  type: integer
                skills:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      category:
                        type: string
                      level:
                        type: string
                stipend:
                  type: object
                  properties:
                    min:
                      type: number
                    max:
                      type: number
                    currency:
                      type: string
                    period:
                      type: string
                      enum: [monthly, hourly, stipend]
                workType:
                  type: string
                  enum: [remote, hybrid, onsite]
                location:
                  type: string
                duration:
                  type: object
                  properties:
                    min:
                      type: integer
                    max:
                      type: integer
                startDate:
                  type: string
                  format: date
                applicationDeadline:
                  type: string
                  format: date
      responses:
        201:
          description: Requirement created

  # Assessments
...
  /communication/chats:
    get:
      tags: [Communication]
      summary: Get user chats
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        200:
          description: Chats retrieved

    post:
      tags: [Communication]
      summary: Create new chat
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - participants
              properties:
                participants:
                  type: array
                  items:
                    type: string
                name:
                  type: string
                isGroupChat:
                  type: boolean
      responses:
        201:
          description: Chat created
  /communication/chats/{chatId}/messages:
    get:
      tags: [Communication]
      summary: Get chat messages
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        200:
          description: Messages retrieved

    post:
      tags: [Communication]
      summary: Send message
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                type:
                  type: string
                  enum: [text, image, file, audio, video]
                attachments:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      url:
                        type: string
                      type:
                        type: string
                      size:
                        type: integer
      responses:
        201:
          description: Message sent